# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Interview Coach API
  version: 1.0.0
  description: Static API spec derived from the current Spring controllers and DTOs.
servers:
  - url: http://localhost:8080
    description: Local backend
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    UUID:
      type: string
      format: uuid
...
    InterviewDirection:
      type: string
      enum: [BACKEND, FRONTEND, DEVOPS]
    InterviewLevel:
      type: string
      enum: [JUNIOR, MIDDLE]
...
    QuestionType:
      type: string
      enum: [TECHNICAL, BEHAVIORAL, GENERAL]
    QuestionStatus:
      type: string
      enum: [ACTIVE, DISABLED]
...
    ProfileQuestionResponse:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        questionId:
          $ref: '#/components/schemas/UUID'
        questionText:
          type: string
        questionType:
          type: string
        orderIndex:
          type: integer
        required:
          type: boolean
...
    ProfileQuestionRequest:
      type: object
      required: [questionId, orderIndex]
      properties:
        questionId:
          $ref: '#/components/schemas/UUID'
        orderIndex:
          type: integer
        required:
          type: boolean
          default: true
    QuestionRequest:
      type: object
      required: [text, questionType]
      properties:
        text:
          type: string
        questionType:
          $ref: '#/components/schemas/QuestionType'
        difficulty:
          $ref: '#/components/schemas/InterviewLevel'
        direction:
          $ref: '#/components/schemas/InterviewDirection'
        status:
          $ref: '#/components/schemas/QuestionStatus'
    QuestionResponse:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        text:
          type: string
        questionType:
          $ref: '#/components/schemas/QuestionType'
        difficulty:
          $ref: '#/components/schemas/InterviewLevel'
        direction:
          $ref: '#/components/schemas/InterviewDirection'
        status:
          $ref: '#/components/schemas/QuestionStatus'
        createdBy:
          $ref: '#/components/schemas/UUID'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    QuestionPageResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/QuestionResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
...
security:
  - bearerAuth: []
paths:
...
  /admin/questions:
    get:
      tags: [Admin Questions]
      summary: List questions
      parameters:
        - name: query
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: direction
          in: query
          schema:
            $ref: '#/components/schemas/InterviewDirection'
        - name: difficulty
          in: query
          schema:
            $ref: '#/components/schemas/InterviewLevel'
        - name: questionType
          in: query
          schema:
            $ref: '#/components/schemas/QuestionType'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/QuestionStatus'
        - name: excludeProfileId
          in: query
          schema:
            $ref: '#/components/schemas/UUID'
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 50
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [updatedAt, createdAt, text]
            default: updatedAt
        - name: sortDir
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Questions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionPageResponse'
    post:
      tags: [Admin Questions]
      summary: Create question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionRequest'
      responses:
        '200':
          description: Question created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionResponse'
  /admin/questions/{questionId}:
    get:
      tags: [Admin Questions]
      summary: Get question by id
      parameters:
        - name: questionId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Question
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionResponse'
    patch:
      tags: [Admin Questions]
      summary: Update question
      parameters:
        - name: questionId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionRequest'
      responses:
        '200':
          description: Question updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionResponse'
    delete:
      tags: [Admin Questions]
      summary: Delete question
      parameters:
        - name: questionId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '204':
          description: Question deleted
...
  /admin/profiles/{profileId}/questions/{linkId}:
    patch:
      tags: [Admin Profile Questions]
      summary: Update profile question link
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: linkId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileQuestionRequest'
      responses:
        '200':
          description: Link updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileQuestionResponse'
    delete:
      tags: [Admin Profile Questions]
      summary: Delete profile question link
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: linkId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '204':
          description: Link deleted